+++ django_rq/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "4.1.0"
+__version__ = "4.1.1"
 
 from .connection_utils import get_connection
 from .decorators import job